Show beacon status on vessel sidebar warning click
