+++ frontend/src/features/backoffice/create_regulation/custom_form/CustomDatePicker.js
@@ -92,18 +92,18 @@ const DatePickerStyled = styled(DatePicker)`
   }
 
   .rs-picker-toggle.rs-btn {
-    border: 1px solid ${props => props.$isrequired ? COLORS.red : COLORS.lightGray}  !important;
-    border-color: ${props => props.$isrequired ? COLORS.red : COLORS.lightGray} !important;
+    border: 1px solid ${props => props.$isrequired ? COLORS.maximumRed : COLORS.lightGray}  !important;
+    border-color: ${props => props.$isrequired ? COLORS.maximumRed : COLORS.lightGray} !important;
     box-sizing: border-box;
     padding: 6px;
   }
   .rs-picker-toggle.rs-btn:focus {
-    border: 1px solid ${props => props.$isrequired ? COLORS.red : COLORS.lightGray}  !important;
-    border-color: ${props => props.$isrequired ? COLORS.red : COLORS.lightGray} !important;
+    border: 1px solid ${props => props.$isrequired ? COLORS.maximumRed : COLORS.lightGray}  !important;
+    border-color: ${props => props.$isrequired ? COLORS.maximumRed : COLORS.lightGray} !important;
   }
   .rs-picker-toggle.rs-btn:hover {
-    border: 1px solid ${props => props.$isrequired ? COLORS.red : COLORS.lightGray}  !important;
-    border-color: ${props => props.$isrequired ? COLORS.red : COLORS.lightGray} !important;
+    border: 1px solid ${props => props.$isrequired ? COLORS.maximumRed : COLORS.lightGray}  !important;
+    border-color: ${props => props.$isrequired ? COLORS.maximumRed : COLORS.lightGray} !important;
   }
 `
