+++ frontend/src/features/vessel_sidebar/controls/Control.js
@@ -255,7 +255,7 @@ const ResumeBoxText = styled.span`
 `
 
 const ResumeBoxNumber = styled.span`
-  background: ${props => props.isRed ? COLORS.red : COLORS.charcoal};
+  background: ${props => props.isRed ? COLORS.maximumRed : COLORS.charcoal};
   color: ${COLORS.gainsboro};
   border-radius: 11px;
   height: 16px;
